Added restriction for admin to delete itself in users/edit view

+++ resources/views/users/edit.blade.php
@@ -67,16 +67,18 @@
                     </div>
                 {!! Form::close() !!}
 
-                <a class="btn btn-danger pull-right" href="" onclick="
-                    event.preventDefault();
-                    if(confirm('Delete user?')) {
-                        document.getElementById('user-delete-form').submit();
-                    }
-                ">
-                    Delete User
-                </a>
-                {!! Form::open(['action' => ['UsersController@destroy', auth()->user()->id], 'method' => 'DELETE', 'id' => 'user-delete-form']) !!}
-                {!! Form::close() !!}
+                @if(auth()->user()->role === 'user' || auth()->user()->role === 'mod')
+                    <a class="btn btn-danger pull-right" href="" onclick="
+                        event.preventDefault();
+                        if(confirm('Delete user?')) {
+                            document.getElementById('user-delete-form').submit();
+                        }
+                    ">
+                        Delete User
+                    </a>
+                    {!! Form::open(['action' => ['UsersController@destroy', auth()->user()->id], 'method' => 'DELETE', 'id' => 'user-delete-form']) !!}
+                    {!! Form::close() !!}
+                @endif
             </div>
         </div>
     </div>
